wait up to 60s for tvheadend to be available on start
When multiple containers for tvheadend, horepg, etc are started in parallel,
for example with docker-compose, horepg may start before tvheadend is listening
for http requests. In this situation the horepg container would exit.

Using 'depends on' ordering in docker-compose is not enough since the
container will not be immediately listening for http requests.

A wait-for script was added to wait up to 60s for tvheadend to respond.